+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Features
 * **Android:** Support for Android's Legacy Mode rendering. [#83](https://github.com/matteobortolazzo/HtmlLabelPlugin/pull/83)
+* **Android:** Support for images. [#49](https://github.com/matteobortolazzo/HtmlLabelPlugin/issues/49)
 
 # 4.0.1